AI agents/project02/pyproject.toml
[project]
name = "project02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.73.0",
    "python-dotenv",
]
